requirements.txt
Django==2.1.7
django-cors-headers==2.4.1
django-extensions==2.1.6
django-imagekit==4.0.2
django-nested-admin==3.1.3
djangorestframework==3.9.2
fastai==1.0.47.post1
ipdb==0.11
ipython==7.3.0
Pillow==5.3.0
